src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/upgrade-candidate-remove-crash-expected.txt

CONSOLE ERROR: line 19: Uncaught TypeError: document.registerElement is not a function
Tests that an element is not upgraded after being destroyed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
